requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
requests==2.31.0
pytest==9.0.2